Reconcile resource requirements for exporter
